improve map interface
