+++ package.json
@@ -14,29 +14,29 @@
     "@headlessui/tailwindcss": "^0.2.0",
     "@heroicons/react": "^2.0.18",
     "@react-google-maps/api": "^2.19.2",
-    "@supabase/ssr": "^0.0.9",
-    "@supabase/supabase-js": "^2.38.4",
-    "@types/node": "20.9.0",
-    "@types/react": "18.2.37",
-    "@types/react-dom": "18.2.15",
+    "@supabase/ssr": "^0.0.10",
+    "@supabase/supabase-js": "^2.38.5",
+    "@types/node": "20.10.0",
+    "@types/react": "18.2.39",
+    "@types/react-dom": "18.2.17",
     "autoprefixer": "10.4.16",
-    "eslint": "8.53.0",
-    "eslint-config-next": "14.0.1",
-    "next": "^14.0.1",
+    "eslint": "8.54.0",
+    "eslint-config-next": "14.0.3",
+    "next": "^14.0.3",
     "postcss": "8.4.31",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-hot-toast": "^2.4.1",
     "tailwind-merge": "^2.0.0",
     "tailwindcss": "3.3.5",
-    "typescript": "5.2.2",
+    "typescript": "5.3.2",
     "zod": "^3.22.4"
   },
   "devDependencies": {
-    "@tailwindcss/forms": "^0.5.6",
+    "@tailwindcss/forms": "^0.5.7",
     "encoding": "^0.1.13",
-    "prettier": "^3.0.3",
-    "prettier-plugin-tailwindcss": "^0.5.6",
-    "supabase": "^1.110.1"
+    "prettier": "^3.1.0",
+    "prettier-plugin-tailwindcss": "^0.5.7",
+    "supabase": "^1.113.3"
   }
 }
