+++ src/app/contact/page.tsx
@@ -1,5 +1,5 @@
-import Button from "@/components/Button";
-import Input from "@/components/Input";
+import Button from "@/components/ui/Button";
+import Input from "@/components/ui/Input";
 
 export default async function ContactPage() {
   return (
