+++ src/app/layout.tsx
@@ -1,8 +1,10 @@
+import type { Metadata } from "next";
+import Image from "next/image";
+
+import "@/app/globals.css";
 import { Auth } from "@/components/Auth";
 import { Navbar } from "@/components/Navbar";
-import type { Metadata } from "next";
-import "./globals.css";
-import Toaster from "./toaster";
+import Toaster from "@/components/toaster";
 
 export const metadata: Metadata = {
   title: "The Local Bard",
@@ -21,11 +23,24 @@ export default async function RootLayout({
 }) {
   return (
     <html lang="en" className="h-full">
-      <body className="flex h-full flex-col overscroll-none bg-[radial-gradient(ellipse_at_top_left,_var(--tw-gradient-stops))] from-zinc-600 via-zinc-900 to-zinc-800 bg-fixed">
+      <body className="flex h-full flex-col overscroll-none bg-fixed">
+        <Image
+          src="/spotlights.png"
+          alt="Spotlights"
+          style={{
+            objectFit: "cover",
+            zIndex: -1,
+            position: "fixed",
+            height: "100vh",
+            width: "100vw",
+          }}
+          height={1080}
+          width={1920}
+        />
         <Navbar>
           <Auth />
         </Navbar>
-        <main className="flex grow">{children}</main>
+        <main className="mt-20 flex grow">{children}</main>
         <Toaster
           position="top-right"
           toastOptions={{
