+++ src/app/reset-password/check-email/page.tsx
@@ -1,4 +1,4 @@
-import Button from "@/components/Button";
+import Button from "@/components/ui/Button";
 import Link from "next/link";
 
 export default function CheckEmailPage() {
