+++ src/components/CreateProductionForm.tsx
@@ -5,10 +5,10 @@ import toast from "react-hot-toast";
 
 import createProduction from "@/actions/createProduction";
 import type { TheaterForNewProduction } from "@/lib/supabase/queries";
-import Button from "./Button";
-import Input from "./Input";
-import Label from "./Label";
-import SubmitButton from "./SubmitButton";
+import Button from "./ui/Button";
+import Input from "./ui/Input";
+import Label from "./ui/Label";
+import SubmitButton from "./ui/SubmitButton";
 
 interface ProductionFormProps {
   theater: TheaterForNewProduction;
