+++ src/components/Navbar.tsx
@@ -2,11 +2,17 @@
 import { Bars3Icon, XMarkIcon } from "@heroicons/react/20/solid";
 import Image from "next/image";
 import React from "react";
+import { usePathname } from "next/navigation";
 
 import { Disclosure } from "@headlessui/react";
-import Button from "./Button";
+import Button from "./ui/Button";
+import { twMerge } from "tailwind-merge";
+import Link from "next/link";
 
 export const Navbar = ({ children }: { children: React.ReactNode }) => {
+  const pathname = usePathname();
+  const isSearchPage = pathname === "/search";
+
   const links = [
     { href: "#", label: "Add a theater" },
     { href: "#", label: "Claim a theater" },
@@ -14,24 +20,40 @@ export const Navbar = ({ children }: { children: React.ReactNode }) => {
   ];
 
   return (
-    <Disclosure as="nav" className="bg-transparent text-zinc-200">
+    <Disclosure
+      as="nav"
+      className={twMerge(
+        "text-zinc-200",
+        isSearchPage
+          ? "fixed top-0 z-50 w-full shrink-0 bg-zinc-900"
+          : "bg-transparent",
+      )}
+    >
       {({ open }) => (
         <>
           {/* Desktop Navbar */}
-          <div className="mx-auto hidden w-full max-w-5xl items-center justify-between p-8 md:flex">
-            <a href="/" className="flex items-center gap-2 font-medium">
+          <div
+            className={twMerge(
+              "mx-auto hidden h-20 w-full max-w-5xl items-center justify-between p-8 md:flex",
+              isSearchPage && "bg-zinc-900",
+            )}
+          >
+            <Link href="/" className="flex items-center gap-2 font-medium">
               <Image width={40} height={40} alt="" src="/logo.jpg" />
               The Local Bard
-            </a>
+            </Link>
             <div className="flex items-center gap-x-4 text-sm">
               {links.map((link) => (
-                <a
+                <Link
                   key={link.label}
                   href={link.href}
-                  className="rounded px-3 py-2 text-zinc-400 hover:bg-zinc-700 hover:text-white"
+                  className={twMerge(
+                    "rounded px-3 py-2 text-zinc-400 hover:text-white",
+                    isSearchPage && "hover:bg-zinc-700",
+                  )}
                 >
                   {link.label}
-                </a>
+                </Link>
               ))}
               {children}
             </div>
