+++ src/components/map/index.tsx
@@ -1,0 +1,3 @@
+import { default as Map } from "./Map";
+
+export default Map;
